Adding Event Notification Intergation Tests
